+++ ggml/src/ggml-hexagon/htp/CMakeLists.txt
@@ -18,7 +18,6 @@ add_library(${HTP_LIB} SHARED
     htp_iface_skel.c
     worker-pool.c
     hex-dma.c
-    hvx-sigmoid.c
     hvx-utils.c
     matmul-ops.c
     binary-ops.c
